requirements/requirements.txt
tensorflow-macos==2.16.2
tensorflow-metal==1.2.0
opencv-python==4.11.0.86
numpy==1.26.4
matplotlib
pandas
pillow